fix(fslib): stop the timeout started by the custom watcher
